+++ loading.js
@@ -1,3 +1,0 @@
-// put it into `.dumi` folder when dumi ready
-import Loading from './.dumi/theme/common/Loading';
-export default Loading;